requirements.txt
aiohttp==3.9.3
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
APScheduler==3.10.4
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
dataclasses-json==0.6.4
DateTime==5.5
distro==1.9.0
exceptiongroup==1.2.0
frozendict==2.4.1
frozenlist==1.4.1
geojson==2.5.0
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-search-results==2.4.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.40
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.40
lxml==5.2.1
marshmallow==3.21.1
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.16.2
orjson==3.10.0
packaging==23.2
pandas==2.2.1
peewee==3.17.1
persiantools==4.0.0
proto-plus==1.23.0
protobuf==4.25.3
pyaes==1.6.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pyowm==3.3.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.1.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
serpapi==0.1.5
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.5
SQLAlchemy==2.0.29
Telethon==1.34.0
tenacity==8.2.3
tiktoken==0.6.0
tornado==6.4
tqdm==4.66.2
types-requests==2.31.0.20240403
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
webencodings==0.5.1
yarl==1.9.4
yfinance==0.2.37
zope.interface==6.2
